requirements.txt
numpy
tensorflow-gpu>=2.2.0
tensorflow-serving-api
opencv-python
matplotlib